uploads: Add authorization check before serving files.
This is a remerge of 473374709e584fce03f3802cfdf719ac28791bbd (after
resolving merge conflicts, updating the tests, adding mypy annotations
etc.), which should now be correct, because we've done the necessary
database migration.

The rebase/remerge work was done by Tim Abbott and Aditya Bansal.

This is an important part of #320.