Add line break to prevent image inline-block'ing.
The image in the page will display next in line to the previous
paragraph which is not how it should look. This adds a <br> to
make the image appear to work as a block element and go down to
the next line without corrupting the <ol> it lives within.