docs: Update the casper docs based on recent debugging.
Having just found that a number of our casper tests were buggy because
they were using casper.waitForSelector and fixed those, this seemed
like a good opportunity to update the docs to recommend the selectors
that more faithfully do what developers expect them to.

While I was at it, I tried to make this a bit better organized, though
I think more work could be done on that front.