font-awesome: Fix font-awesome 4.7.0 upgrade in a hacky way.
What actually has been done below is to just copy the css class defination
from the latest font-awesome css to be here and since the rest of the stuff
in /third/fontawesome was updated in e4fa4b0770c4c51db1023742338166af5961c186
we should be able to use this safely until we update all font-awesome class
usage in templates.

Fixes #4302.