README: Update discussion of mobile apps and of GSoC.
Now that GSoC 2017 is underway, there's no need to say much
about the application process.

That section is now very short and doesn't feel like it
really fits with the other sections around.  But the remaining
content seems like interesting stuff that's nice to have
in the README so that e.g. it's linked from the repo's
front page on GitHub.  Leaving further refactoring to
another day.