recent_senders: Fix exception in certain typeahead flows.
It appears that a regression introduced in
863ed701b2a3acb0051ed1d2b52fe4d5f796a360 caused undefined to be passed
as the subject to the recent_senders library much more often; this
fixing that, and makes the library handle such cases reasonably
without an exception regardless.

This was causing a huge number of "Tried to call a Dict method with an
undefined key." exceptions.