Fix deleting stream from the sidebar.
When we were deleting a stream from the sidebar using the
stream/delete event, we were getting tracebacks due to this sequence
of operations:

    * remove id from stream_list.stream_sidebar
    * rebuild stream list
    * remove sub from stream_data

This fixes the bug by calling stream_data.delete_sub() first.

Deletions are tricky if you do things out of order.  We can probably
prevent tracebacks by having a deleted flag, but that can just cause
different problems.

Last commit tweaked by tabbott to fix a small bug in handling the case
where the user was not subscribed to the delete stream.