Remove the scaling and box shadow on hovered price options.
This removes the scaling feature where a plan that a user
hovers over would increase in size and have a box shadow.

This used to be here because there was a clickable region that
would pop open an overlay that would have more plan information,
but that feature no longer exists.