i18n: Update Japanese translations.
This comes from, in the dev environment:
  ./manage.py makemessages -l en -l ja
  tx push -s -l ja
  tx pull -f -l ja --mode=developer

The makemessages and `tx push` provide Transifex with the updated set
of strings and metadata.  I'm not entirely sure why that's necessary,
but without it a lone `tx pull` left the server crashing with numerous
zerver.models.DoesNotExist exceptions.

(So to be precise, there was another `tx pull` at the start, just like
the final one.  That shouldn't matter for the result, though.)